+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/TomatoCropBlock.java
@@ -3,34 +3,24 @@ package com.christofmeg.brutalharvest.common.block;
 import com.christofmeg.brutalharvest.common.init.ItemRegistry;
 import com.christofmeg.brutalharvest.common.item.KnifeItem;
 import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.ItemLike;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.CropBlock;
-import net.minecraft.world.level.block.FarmBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraft.world.level.gameevent.GameEvent;
-import net.minecraft.world.level.storage.loot.LootParams;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
 
 public class TomatoCropBlock extends BaseCropBlock {
 
@@ -49,66 +39,6 @@ public class TomatoCropBlock extends BaseCropBlock {
 
     public TomatoCropBlock(Properties properties) {
         super(properties);
-        this.registerDefaultState(this.stateDefinition.any().setValue(this.getAgeProperty(), 0));
-    }
-
-    @Override
-    public @NotNull VoxelShape getShape(@NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
-        return SHAPE_BY_AGE[this.getAge(state)];
-    }
-
-    @Override
-    public void randomTick(@NotNull BlockState state, @NotNull ServerLevel $$1, @NotNull BlockPos $$2, @NotNull RandomSource $$3) {
-        int age = state.getValue(AGE);
-        if (age < 8 && $$3.nextInt(5) == 0 && $$1.getRawBrightness($$2.above(), 0) >= 9) {
-            BlockState $$5 = state;
-            if (age == 7) {
-                if ($$3.nextInt(20) == 0) { // 5% chance
-                    $$5 = state.setValue(AGE, age + 1);
-                }
-            } else {
-                $$5 = state.setValue(AGE, age + 1);
-            }
-            $$1.setBlock($$2, $$5, 2);
-            $$1.gameEvent(GameEvent.BLOCK_CHANGE, $$2, GameEvent.Context.of($$5));
-        }
-    }
-
-    @Override
-    public boolean isRandomlyTicking(@NotNull BlockState pState) {
-        return this.getAge(pState) < this.getMaxAge() + 1;
-    }
-
-    @Override
-    public boolean canSurvive(@NotNull BlockState state, @NotNull LevelReader levelReader, BlockPos pos) {
-        BlockPos below = pos.below();
-        return this.mayPlaceOn(levelReader.getBlockState(below), levelReader, below);
-    }
-
-    @Override
-    public int getMaxAge() {
-        return 7;
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public @NotNull List<ItemStack> getDrops(@NotNull BlockState state, LootParams.@NotNull Builder builder) {
-        List<ItemStack> dropsOriginal = super.getDrops(state, builder);
-        int randomTomatoes = 3 + builder.getLevel().random.nextInt(4);
-        int randomRotten = builder.getLevel().random.nextInt(2);
-        if (!dropsOriginal.isEmpty()) {
-            return dropsOriginal;
-        }
-        return
-                state.getValue(AGE) == 4 || state.getValue(AGE) == 5 || state.getValue(AGE) == 6 ?  List.of(new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                        state.getValue(AGE) == 7 ? List.of(new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                                state.getValue(AGE) == 8 ? List.of(new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), 1 + randomRotten), new ItemStack(this, randomRotten)) :
-                                        List.of(new ItemStack(this, 1));
-    }
-
-    @Override
-    protected @NotNull ItemLike getBaseSeedId() {
-        return ItemRegistry.TOMATO_SEEDS.get();
     }
 
     @Override
@@ -118,42 +48,56 @@ public class TomatoCropBlock extends BaseCropBlock {
     }
 
     @Override
-    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
-        builder.add(AGE);
+    public int getMaxAge() {
+        return 7;
     }
 
     @Override
-    public boolean isValidBonemealTarget(@NotNull LevelReader levelReader, @NotNull BlockPos pos, @NotNull BlockState state, boolean $$3) {
-        return state.getValue(AGE) < 6;
+    protected @NotNull ItemLike getBaseSeedId() {
+        return ItemRegistry.TOMATO_SEEDS.get();
     }
 
-    @SuppressWarnings("deprecation")
+    @Override
+    protected ItemStack getBaseItemStack() {
+        return ItemRegistry.TOMATO.get().getDefaultInstance();
+    }
+
+    @Override
+    protected int getAgeAfterKnife() {
+        return 3;
+    }
+
+    @Override
+    public @NotNull VoxelShape getShape(@NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos pos, @NotNull CollisionContext context) {
+        return SHAPE_BY_AGE[this.getAge(state)];
+    }
+
     @Override
     public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
-        int age = state.getValue(AGE);
-        boolean reachedTomatoAge = age == 4;
-        if (!reachedTomatoAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
-            return InteractionResult.PASS;
-        } else if (age > 3) {
-            ItemStack stack = player.getItemInHand(interactionHand);
-            if (stack.getItem() instanceof KnifeItem) {
-                int randomTomatoes = 3 + level.random.nextInt(4);
-                int randomRotten = 1 + level.random.nextInt(2);
-                popResource(level, pos,
-                        age == 4 || age == 5 || age == 6 ? new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes) :
-                                age == 7 ? new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes) :
-                                        age == 8 ? new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomRotten) : ItemStack.EMPTY);
-                level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
-                BlockState newBlockState = state.setValue(AGE, 3);
-                if (age == 8) {
-                    level.setBlock(pos, Blocks.AIR.defaultBlockState(), 2);
-                } else {
-                    level.setBlock(pos, newBlockState, 2);
-                }
-                level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, newBlockState));
-                stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
-                return InteractionResult.sidedSuccess(level.isClientSide);
+        int age = this.getAge(state);
+        boolean unripeAge = age == 4 || age == 5 || age == 6;
+        boolean matureAge = age == this.getMaxAge();
+        boolean deadAge = age == this.getMaxAge() + 1;
+        ItemStack stack = player.getItemInHand(interactionHand);
+        if (stack.getItem() instanceof KnifeItem && !level.isClientSide) {
+            if (!unripeAge && !matureAge && !deadAge) {
+                return InteractionResult.sidedSuccess(false);
             }
+            if (unripeAge) {
+                state = this.getStateForAge(getAgeAfterKnife());
+                popResource(level, pos, new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), 3 + level.random.nextInt(4)));
+            } else if (matureAge) {
+                state = this.getStateForAge(getAgeAfterKnife());
+                popResource(level, pos, new ItemStack(this.getBaseItemStack().getItem(), 3 + level.random.nextInt(4)));
+            } else {
+                state = Blocks.AIR.defaultBlockState();
+                popResource(level, pos, new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), 3 + level.random.nextInt(4)));
+            }
+            level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
+            level.setBlock(pos, state, 2);
+            level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, state));
+            stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
+            return InteractionResult.sidedSuccess(false);
         }
         return super.use(state, level, pos, player, interactionHand, blockHitResult);
     }