+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/data/loot/BrutalBlockLootTables.java
@@ -1,9 +1,6 @@
 package com.christofmeg.brutalharvest.common.data.loot;
 
-import com.christofmeg.brutalharvest.common.block.CottonCropBlock;
-import com.christofmeg.brutalharvest.common.block.LettuceCropBlock;
-import com.christofmeg.brutalharvest.common.block.StrawberryCropBlock;
-import com.christofmeg.brutalharvest.common.block.SugarBeetCropBlock;
+import com.christofmeg.brutalharvest.common.block.*;
 import com.christofmeg.brutalharvest.common.init.BlockRegistry;
 import com.christofmeg.brutalharvest.common.init.ItemRegistry;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
@@ -35,6 +32,30 @@ public class BrutalBlockLootTables extends BlockLootSubProvider {
     protected void generate() {
 
         //TODO block loot to Corn, Tomato
+
+        this.add(BlockRegistry.TOMATO.get(), createGenericUnripeRottenCropDrops(
+                BlockRegistry.TOMATO.get(),
+                ItemRegistry.UNRIPE_TOMATO.get(), 3.0F, 6.0F,
+                ItemRegistry.TOMATO.get(), 3.0F, 6.0F,
+                ItemRegistry.ROTTEN_TOMATO.get(), 3.0F, 6.0F,
+                ItemRegistry.TOMATO_SEEDS.get(), 1.0F, 2.0F,
+                LootItemBlockStatePropertyCondition
+                        .hasBlockStateProperties(BlockRegistry.TOMATO.get())
+                        .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 4))
+                        .or(LootItemBlockStatePropertyCondition
+                                .hasBlockStateProperties(BlockRegistry.TOMATO.get())
+                                .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 5)))
+                        .or(LootItemBlockStatePropertyCondition
+                                .hasBlockStateProperties(BlockRegistry.TOMATO.get())
+                                .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 6)))
+                ,
+                LootItemBlockStatePropertyCondition
+                        .hasBlockStateProperties(BlockRegistry.TOMATO.get())
+                        .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 7)),
+                LootItemBlockStatePropertyCondition
+                        .hasBlockStateProperties(BlockRegistry.TOMATO.get())
+                        .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 8))
+        ));
 
         this.add(BlockRegistry.LETTUCE.get(), createLettuceCropDrops(
                 BlockRegistry.LETTUCE.get(),
@@ -72,10 +93,14 @@ public class BrutalBlockLootTables extends BlockLootSubProvider {
                         .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(CottonCropBlock.AGE, 6))
         ));
 
-        this.add(BlockRegistry.STRAWBERRY.get(), createGenericCropDrops(
+        this.add(BlockRegistry.STRAWBERRY.get(), createGenericUnripeCropDrops(
                 BlockRegistry.STRAWBERRY.get(),
+                ItemRegistry.UNRIPE_STRAWBERRY.get(), 1.0F, 4.0F,
                 ItemRegistry.STRAWBERRY.get(), 1.0F, 4.0F,
                 ItemRegistry.STRAWBERRY_SEEDS.get(), 1.0F, 2.0F,
+                LootItemBlockStatePropertyCondition
+                        .hasBlockStateProperties(BlockRegistry.STRAWBERRY.get())
+                        .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 5)),
                 LootItemBlockStatePropertyCondition
                         .hasBlockStateProperties(BlockRegistry.STRAWBERRY.get())
                         .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(StrawberryCropBlock.AGE, 6)),
@@ -90,8 +115,9 @@ public class BrutalBlockLootTables extends BlockLootSubProvider {
     protected @NotNull Iterable<Block> getKnownBlocks() {
         return BlockRegistry.BLOCKS.getEntries().stream().map(RegistryObject::get)
                 .filter(block ->
-                        block instanceof SugarBeetCropBlock ||
+                        block instanceof TomatoCropBlock ||
                         block instanceof LettuceCropBlock ||
+                        block instanceof SugarBeetCropBlock ||
                         block instanceof CottonCropBlock ||
                         block instanceof StrawberryCropBlock
                 )
@@ -162,6 +188,68 @@ public class BrutalBlockLootTables extends BlockLootSubProvider {
                                 .when(LootItemRandomChanceCondition.randomChance(0.25F))))
                 .withPool(LootPool.lootPool()
                         .when(matureCondition.invert().and(deadCondition.invert()))
+                        .add(LootItem.lootTableItem(pSeedsItem))
+                )
+        );
+    }
+
+    protected LootTable.Builder createGenericUnripeCropDrops(Block pCropBlock, Item unripeCropItem, float minUnripeCropItem, float maxUnripeCropItem, Item pGrownCropItem, float minCropItem, float maxCropItem, Item pSeedsItem, float minSeedItem, float maxSeedItem, LootItemCondition.Builder unripeConditon, LootItemCondition.Builder matureCondition, LootItemCondition.Builder deadCondition) {
+        return this.applyExplosionDecay(pCropBlock, LootTable.lootTable()
+                .withPool(LootPool.lootPool()
+                        .when(unripeConditon)
+                        .add(LootItem.lootTableItem(unripeCropItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(minUnripeCropItem, maxUnripeCropItem)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(matureCondition)
+                        .add(LootItem.lootTableItem(pGrownCropItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(minCropItem, maxCropItem)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(matureCondition)
+                        .add(LootItem.lootTableItem(pSeedsItem)
+                                .apply(SetItemCountFunction.setCount(UniformGenerator.between(minSeedItem, maxSeedItem))))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(deadCondition)
+                        .add(LootItem.lootTableItem(pSeedsItem)
+                                .when(LootItemRandomChanceCondition.randomChance(0.25F))))
+                .withPool(LootPool.lootPool()
+                        .when(unripeConditon.invert().and(matureCondition.invert().and(deadCondition.invert())))
+                        .add(LootItem.lootTableItem(pSeedsItem))
+                )
+        );
+    }
+
+    protected LootTable.Builder createGenericUnripeRottenCropDrops(Block pCropBlock, Item unripeCropItem, float minUnripeCropItem, float maxUnripeCropItem, Item pGrownCropItem, float minCropItem, float maxCropItem, Item rottenCropItem, float minRottenCropItem, float maxRottenCropItem, Item pSeedsItem, float minSeedItem, float maxSeedItem, LootItemCondition.Builder unripeConditon, LootItemCondition.Builder matureCondition, LootItemCondition.Builder deadCondition) {
+        return this.applyExplosionDecay(pCropBlock, LootTable.lootTable()
+                .withPool(LootPool.lootPool()
+                        .when(unripeConditon)
+                        .add(LootItem.lootTableItem(unripeCropItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(minUnripeCropItem, maxUnripeCropItem)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(matureCondition)
+                        .add(LootItem.lootTableItem(pGrownCropItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(minCropItem, maxCropItem)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(matureCondition)
+                        .add(LootItem.lootTableItem(pSeedsItem)
+                                .apply(SetItemCountFunction.setCount(UniformGenerator.between(minSeedItem, maxSeedItem))))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(deadCondition)
+                        .add(LootItem.lootTableItem(rottenCropItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(minRottenCropItem, maxRottenCropItem)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(deadCondition)
+                        .add(LootItem.lootTableItem(pSeedsItem))
+                        .apply(SetItemCountFunction.setCount(UniformGenerator.between(0, 1)))
+                )
+                .withPool(LootPool.lootPool()
+                        .when(unripeConditon.invert().and(matureCondition.invert().and(deadCondition.invert())))
                         .add(LootItem.lootTableItem(pSeedsItem))
                 )
         );